Move authenticate method to the rpc util
