+++ rpc-server.py
@@ -3,26 +3,15 @@
 import os
 import bcrypt
 
-from functools import wraps
 from flask import Flask, session, request
 from flask_jsonrpc import JSONRPC, _parse_sig
-from flask_jsonrpc.exceptions import InvalidCredentialsError
 from firebase_token_generator import create_token
 from inspect import getargspec
 
 from cabinet import Cabinet
+from rpc import authenticate
 
 app = Flask(__name__)
-
-
-def authenticate(f, f_check_auth):
-    @wraps(f)
-    def _f(*args, **kwargs):
-        token = request.headers.get('token')
-        if not f_check_auth(token):
-            raise InvalidCredentialsError()
-        return f(*args, **kwargs)
-    return _f
 
 
 class ExtendedJSONRPC(JSONRPC):
