+++ rpc/__init__.py
@@ -1,0 +1,3 @@
+from .util import authenticate
+
+__all__ = ['authenticate']
